+++ administrator/components/com_loginguard/src/View/Tools/HtmlView.php
@@ -4,8 +4,8 @@ namespace LoginGuard\Component\LoginGuard\Administrator\View\Tools;
 
 defined('_JEXEC') or die;
 
+use Joomla\CMS\HTML\Helpers\Sidebar;
 use Joomla\CMS\MVC\View\HtmlView as BaseHtmlView;
-use Joomla\CMS\Session\Session;
 use Joomla\CMS\Toolbar\ToolbarHelper;
 use LoginGuard\Component\LoginGuard\Administrator\Helper\LoginGuardHelper;
 
@@ -23,12 +23,12 @@ final class HtmlView extends BaseHtmlView
 
         $this->actions = LoginGuardHelper::getActions();
         LoginGuardHelper::addSubmenu('tools');
-        $this->sidebar = \Joomla\CMS\HTML\HTMLHelper::_('sidebar.render');
+        $this->sidebar = Sidebar::render();
 
         ToolbarHelper::title('LoginGuard: Tools', 'wrench');
 
         if ($this->actions->get('loginguard.export')) {
-            ToolbarHelper::link('index.php?option=com_loginguard&task=attempts.export&' . Session::getFormToken() . '=1', 'COM_LOGINGUARD_TOOLBAR_EXPORT', 'download');
+            ToolbarHelper::custom('attempts.export', 'download', '', 'COM_LOGINGUARD_TOOLBAR_EXPORT', false);
         }
 
         parent::display($tpl);