+++ administrator/components/com_loginguard/tmpl/tools/default.php
@@ -2,6 +2,7 @@
 
 defined('_JEXEC') or die;
 
+use Joomla\CMS\HTML\HTMLHelper;
 use Joomla\CMS\Language\Text;
 use Joomla\CMS\Router\Route;
 
@@ -27,5 +28,7 @@ use Joomla\CMS\Router\Route;
                 </div>
             </div>
         </div>
+        <input type="hidden" name="task" value="">
+        <?php echo HTMLHelper::_('form.token'); ?>
     </div>
 </form>